% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRDA_0.R
\name{cohen_d}
\alias{cohen_d}
\title{Compute Cohen's d}
\usage{
cohen_d(x1, x2)
}
\arguments{
\item{x1}{= numerical vector with the observed values of the first group}

\item{x2}{= numerical vector with the observed values of the second group}
}
\value{
Returns a list with the following objects: \code{d} (estimated Cohen's \eqn{d}), 
\code{tstat} (\eqn{t} value), \code{pval} (\eqn{p}-value).
}
\description{
Computes Cohen's \eqn{d} effect size statistics with the relative \eqn{t}-statistics and \eqn{p}-value.
}
\details{
Cohen's \eqn{d} is computed considering the pooled standard deviation.

\eqn{t}-statistics and relative \eqn{p}-value are computed considering two-tailed \eqn{t}-test for independent samples with equal variance.
}
\examples{
x1 <- c( 30.02, 29.99, 30.11, 29.97, 30.01, 20.99 )
x2 <- c( 29.89, 29.93, 29.72, 29.98, 30.02, 29.98, 50 )
cohen_d( x1, x2 )
}
\author{
Massimiliano Pastore
}
