% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRDA_0.R
\name{design_analysis}
\alias{design_analysis}
\title{Design Analysis}
\usage{
design_analysis(d, n = NULL, power = NULL, sig.level = 0.05,
  B = 10000, rangen = c(2, 1000), tol = 0.005)
}
\arguments{
\item{d}{= plausible effect size (Cohen's \eqn{d})}

\item{n}{= number of observations (per group)}

\item{power}{= required power of the test}

\item{sig.level}{= significance level}

\item{B}{= number of replicates}

\item{rangen}{= range of \eqn{n} (per group), only if \code{n=NULL}}

\item{tol}{= tolerance in estimation}
}
\value{
Returns a list with the following objects: \code{d} (effect size),
\code{power}, \code{n} (sample size), \code{typeS} (type S error), 
\code{typeM} (type M error).
}
\description{
Performs a prospective or retrospective design analysis for the comparison between two independent groups,
given a plausible value of effect size (Cohen's \eqn{d}). If \code{power} is defined,
the function estimates the sample size (per group) needed to reach that power. Instead, if 
the number of observations (per group) is defined (\code{n}), the function estimates 
the power level, type M and type S errors given that sample size.
}
\details{
The function can be used to compute the power (\code{power})
or the sample size needed (\code{n}). In the first case, users have to set \code{power=NULL} 
and provide values for \code{d} and \code{n}. In the second case, users have to set 
\code{n=NULL} and provide values for \code{d} and \code{power}.

Users can also define through \code{rangen}, the minimum and maximum values
for the range number of observations (per group). When the function evaluates the sample size 
(per group) needed to reach a given power, only values within this range are considered. Default
settings are \code{rangen = c( 2, 1000 )}.
}
\note{
Plausible \eqn{d} is a fixed value, \code{n} is the sample size per group.
}
\examples{
design_analysis( n = 10, d = .5 )
design_analysis( d = .5, power = .10 )
design_analysis( d = .2, power = .90, rangen = c(5,50) )
design_analysis( d = .2, power = .90, rangen = c(50,100) )
}
\references{
Altoè, G., Bertoldo, G., Zandonella Callegher, C., Toffalini E., Calcagnì, A., Finos, L., Pastore, M. (2019).
Enhancing statistical inference in psychological research via prospective and retrospective design analysis. \url{https://arxiv.org/abs/1909.13773} (Submitted to \emph{Frontiers in Psychology}).

Cohen, J. (1988). \emph{Statistical power analysis}. Academic Press, San Diego, CA.

Gelman, A., Carlin, J. (2014). Beyond power calculations: Assessing type S (sign) and type M
(magnitude) errors. \emph{Perspectives on Psychological Science}, 9(6), 641--651. \url{https://doi.org/10.1177/1745691614551642}
}
\seealso{
\code{\link{power.t.test}}
}
\author{
Massimiliano Pastore, Gianmarco Altoè
}
