% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRDA_0.R
\name{design_analysis_power}
\alias{design_analysis_power}
\title{Design Analysis: estimated power}
\usage{
design_analysis_power(n, d, sig.level = 0.05, B = 10000, tol = 0.005)
}
\arguments{
\item{n}{= number of observations (per group)}

\item{d}{= plausible effect size (Cohen's \eqn{d})}

\item{sig.level}{= significance level}

\item{B}{= number of replicates}

\item{tol}{= tolerance}
}
\description{
Internal function. In a comparison between two independent groups,
the function estimates the power level, type M and type S errors given the specified 
sample size per group (\code{n}) and a plausible value of effect size (Cohen's \eqn{d}).
}
\note{
Called from function \code{\link{design_analysis}}.
}
\author{
Massimiliano Pastore, Gianmarco Altoe
}
