% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRDA_0.R
\name{sim_estimate}
\alias{sim_estimate}
\title{Internal function}
\usage{
sim_estimate(sim_d, target_d = sim_d, n1, n2 = n1, sig.level = 0.05,
  B = 10000)
}
\arguments{
\item{sim_d}{= simulated effect size (Cohen's \eqn{d})}

\item{target_d}{= target effect size (Cohen's \eqn{d})}

\item{n1}{= sample size group 1}

\item{n2}{= sample size group 2}

\item{sig.level}{= significance level}

\item{B}{= number of replicates}
}
\value{
Returns a \code{data.frame} with the results of the simulation and the computed values of 
\code{power}, \code{typeS}, and \code{typeM} errors.
}
\description{
Internal function. In a comparison between two independent groups,
the function computes through \code{B} simulations the expected power, 
Type-S and Type-M errors, given a plausible effect size (\code{target_d}) and the sample size of each group.
}
\note{
Called from \code{\link{design_est}}. For \code{design_analysis}: \code{target_d} is equal to \code{sim_d}.
}
\author{
Massimiliano Pastore, Gianmarco Altoè
}
